backend/database/init_db.py
from sqlalchemy import create_engine
from backend.database.models import Base
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Get database URL from environment variable or use default
DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./client_dashboard.db")

# Create engine
engine = create_engine(DATABASE_URL)

# Create all tables
def init_db():
    Base.metadata.create_all(bind=engine)
    print("Database tables created successfully!")

if __name__ == "__main__":
    init_db() 