requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
aiohttp==3.8.1
PyJWT==2.3.0
black==22.3.0
pytest==7.0.1
pytest-asyncio==0.18.3
httpx==0.23.0
pydantic==1.9.0 